+++ server/.gitignore
@@ -4,3 +4,4 @@ node_modules
 keys
 uploads
 temp
+google-credentials.json
